+++ realm/realm-library/src/main/java/io/realm/log/RealmLog.java
@@ -18,6 +18,8 @@ package io.realm.log;
 
 import android.util.Log;
 
+import java.util.Locale;
+
 
 /**
  * Global logger used by all Realm components.
@@ -275,7 +277,7 @@ public final class RealmLog {
 
         StringBuilder stringBuilder = new StringBuilder();
         if (args != null && args.length > 0) {
-            message = String.format(message, args);
+            message = String.format(Locale.US, message, args);
         }
         if (throwable != null) {
             stringBuilder.append(Log.getStackTraceString(throwable));